requirements.txt
numpy
pandas
scikit-learn
sentence-transformers
librosa
opencv-python
faiss-cpu
joblib
flask